requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.2
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.15.4
flatbuffers==24.12.23
fonttools==4.53.1
fsspec==2024.6.1
humanfriendly==10.0
idna==3.8
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==2.1.0
onnx==1.17.0
onnxruntime==1.20.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
protobuf==5.29.2
pyparsing==3.1.4
pyserial==3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
sentencepiece==0.2.0
setuptools==73.0.1
six==1.16.0
sympy==1.13.2
tokenizers==0.19.1
typing_extensions==4.12.2
urllib3==2.2.2
